Add Docker image with export infrastructure.
This commits updates the Dockerfile to add a new structurizr-export on top of the original
structurizr-lite image. The image is configured with puppeteer, headless google chrome, the
puppeteer structurizr export repository and 2 scripts to do the actual exports.

With this image, one can run

```
docker run --rm -v .:/workspace -it structurizr-export:latest export-diagrams workspace build svg
```

to export the diagrams to svg. There's also an export-documentation script.